// src/store/action-types.ts
export enum ActionTypes {
  SET_PLAYER_OPEN_ACTION = "SET_PLAYER_OPEN_ACTION",
  PAUSE_ACTION = "PAUSE_ACTION",
  PLAY_ACTION = "PLAY_ACTION",
  TOGGLE_PLAYING_ACTION = "TOGGLE_PLAYING_ACTION",
  SEEK_ACTION = "SEEK_ACTION",
  NEXT_ACTION = "NEXT_ACTION",
  PREV_ACTION = "PREV_ACTION",
  FAV_ACTION = "FAV_ACTION",
  LOGGED_IN_ACTION = "LOGGED_IN_ACTION",
  LOGOUT_ACTION = "LOGOUT_ACTION",
  ADD_SONGS_ACTION = "ADD_SONGS_ACTION",
  ADD_SONGS_IF_NOT_ACTION = "ADD_SONGS_IF_NOT_ACTION",
  PERCENT_PLAYED_ACTION = "PERCENT_PLAYED_ACTION",
  AUDIO_TIME_ACTION = "AUDIO_TIME_ACTION"
}